android: Extract BaseCobaltActivity from CobaltActivity
This is a precursor PR to ease the review of the AOSP work with no functional
changes.

Move the content-free pieces of CobaltActivity into a new BaseCobaltActivity
base class shared with AOSP.

CobaltActivity now extends BaseCobaltActivity and keeps the Chromium
content/shell wiring. This lets AOSP use the base class instead of creating a
new activity from scratch.

Bug: 495203133
